Handle visited and relative URLs
I have some issues with URLs such as "#" or those missing schemas, but I'm pretty weirded out about the fact that hyper does not provide a way to do this.

Also, I clone the Uri struct 3 times, which I'm sure is fine from a memory standpoint, but for some reason in Rust I hate allocation.